requirements.txt
Django==5.2.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
mysqlclient==2.2.7
python-dotenv==1.0.1
scikit-learn==1.7.2
numpy==2.2.6
scipy==1.15.3
joblib==1.3.2
pandas==2.2.2
Pillow==10.3.0